added mongoDB to implement persistence see issue 4
MongoDB added to project to serve as a database for all emails received

At the moment while emails are read and emitted in FileWatcher they are
now also stored in MongoDB under the emails collection.
A new field is added to the mail object which is used as a key to
retrieve all emails for a minus mail email address.